requirements.txt
requests==2.31.0
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1
click==8.1.7
